Moved out OrganizationFeature class and some functions
