+++ resqpy/organize/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['OrganizationFeature',]
+
+from .OrganizationFeature import OrganizationFeature
