+++ tests/test_model.py
@@ -2,7 +2,6 @@ import os
 
 import numpy as np
 import pytest
-from numpy.lib.arraysetops import isin
 
 import resqpy.crs as rqc
 import resqpy.model as rq
@@ -61,7 +60,7 @@ def test_model_iterators(example_model_with_well):
 
 def test_model_iterate_objects(example_model_with_well):
 
-    from resqpy.organize import WellboreFeature, WellboreInterpretation
+    from resqpy.organize.organize_old import WellboreFeature, WellboreInterpretation
 
     model, well_interp, _, _ = example_model_with_well
