+++ tests/test_resqpy.py
@@ -1,5 +1,3 @@
-import pytest
-
 from resqpy import model
 
 
@@ -9,8 +7,5 @@ def test_empty_model():
 
 
 def test_all_imports():
-    from resqpy import (crs, derived_model, fault, grid, grid_surface, lines, organize, property, rq_import, surface,
-                        time_series, well)
-
     #    from resqpy.olio import *
     return
