+++ tests/test_strata.py
@@ -3,13 +3,12 @@
 import os
 
 import numpy as np
-import pandas as pd
 
 import resqpy.crs as rqc
 import resqpy.grid as grr
 import resqpy.model as rq
 import resqpy.olio.uuid as bu
-import resqpy.organize as rqo
+import resqpy.organize.organize_old as rqo
 import resqpy.strata as strata
 
 # the following creates a fully related set of objects, saves to persistent storage, and retrieves
